# .schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "runtime",
    "description": "The spiced runtime",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.2.0-unstable"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server. Configure with `--http`."
    }
  ],
  "paths": {
...
    "/v1/embeddings": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Create Embeddings",
        "description": "Creates an embedding vector representing the input text.\n\nGet a vector representation of a given input that can be easily consumed by machine learning models and algorithms.",
        "operationId": "post_embeddings",
        "requestBody": {
          "description": "Embedding creation request parameters",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmbeddingRequest"
              },
              "example": {
                "input": "The food was delicious and the waiter...",
                "model": "text-embedding-ada-002",
                "encoding_format": "float"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Embedding created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateEmbeddingResponse"
                },
                "example": {
                  "object": "list",
                  "data": [
                    {
                      "object": "embedding",
                      "embedding": [
                        0.0023064255,
                        -0.009327292,
                        -0.0028842222
                      ],
                      "index": 0
                    }
                  ],
                  "model": "text-embedding-ada-002",
                  "usage": {
                    "prompt_tokens": 8,
                    "total_tokens": 8
                  }
                }
              }
            }
          },
          "404": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": {
                  "error": "model not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": {
                  "error": "Unexpected internal server error occurred"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/iceberg/namespaces": {
      "get": {
        "tags": [
          "Iceberg"
        ],
        "summary": "List Iceberg namespaces",
        "description": "This endpoint retrieves namespaces available in the Iceberg catalog.\nIf a `parent` namespace is provided, it will list the child namespaces under the specified parent.",
        "operationId": "get_iceberg_namespaces",
        "parameters": [
          {
            "name": "parent",
            "in": "query",
            "description": "The parent namespace from which to retrieve child namespaces.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Namespace"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Namespaces retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespacesResponse"
                },
                "example": {
                  "namespaces": [
                    {
                      "parts": [
                        "catalog_a"
                      ]
                    },
                    {
                      "parts": [
                        "catalog_b",
                        "schema_1"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcebergResponseError"
                },
                "example": {
                  "error": {
                    "message": "Invalid namespace request",
                    "type": "BadRequestException",
                    "code": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcebergResponseError"
                },
                "example": {
                  "error": {
                    "message": "Namespace provided does not exist",
                    "type": "NoSuchNamespaceException",
                    "code": 404
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IcebergResponseError"
                },
                "example": {
                  "error": {
                    "message": "Internal Server Error: DF_SCHEMA_NOT_FOUND",
                    "type": "InternalServerError",
                    "code": 500
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEmbeddingRequest": {
        "type": "object",
        "required": [
          "model",
          "input"
        ],
        "properties": {
          "dimensions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The number of dimensions the resulting output embeddings should have. Only supported in `text-embedding-3` and later models.",
            "minimum": 0
          },
          "encoding_format": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EncodingFormat",
                "description": "The format to return the embeddings in. Can be either `float` or [`base64`](https://pypi.org/project/pybase64/). Defaults to float"
              }
            ]
          },
          "input": {
            "$ref": "#/components/schemas/EmbeddingInput",
            "description": "Input text to embed, encoded as a string or array of tokens. To embed multiple inputs in a single request, pass an array of strings or array of token arrays. The input must not exceed the max input tokens for the model (8192 tokens for `text-embedding-ada-002`), cannot be an empty string, and any array must be 2048 dimensions or less. [Example Python code](https://cookbook.openai.com/examples/how_to_count_tokens_with_tiktoken) for counting tokens."
          },
          "model": {
            "type": "string",
            "description": "ID of the model to use. You can use the\n[List models](https://platform.openai.com/docs/api-reference/models/list)\nAPI to see all of your available models, or see our\n[Model overview](https://platform.openai.com/docs/models/overview)\nfor descriptions of them."
          },
          "user": {
            "type": [
              "string",
              "null"
            ],
            "description": "A unique identifier representing your end-user, which will help OpenAI\n to monitor and detect abuse. [Learn more](https://platform.openai.com/docs/usage-policies/end-user-ids)."
          }
        }
      },
      "CreateEmbeddingResponse": {
        "type": "object",
        "required": [
          "object",
          "model",
          "data",
          "usage"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Embedding"
            },
            "description": "The list of embeddings generated by the model."
          },
          "model": {
            "type": "string",
            "description": "The name of the model used to generate the embedding."
          },
          "object": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/EmbeddingUsage",
            "description": "The usage information for the request."
          }
        }
      },
...
      "Embedding": {
        "type": "object",
        "description": "Represents an embedding vector returned by embedding endpoint.",
        "required": [
          "index",
          "object",
          "embedding"
        ],
        "properties": {
          "embedding": {
            "$ref": "#/components/schemas/EmbeddingVector",
            "description": "The embedding vector, which is a list of floats. The length of vector\ndepends on the model as listed in the [embedding guide](https://platform.openai.com/docs/guides/embeddings)."
          },
          "index": {
            "type": "integer",
            "format": "int32",
            "description": "The index of the embedding in the list of embeddings.",
            "minimum": 0
          },
          "object": {
            "type": "string",
            "description": "The object type, which is always \"embedding\"."
          }
        }
      },
      "EmbeddingInput": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "type": "array",
            "items": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32",
                "minimum": 0
              }
            }
          }
        ]
      },
      "EmbeddingUsage": {
        "type": "object",
        "required": [
          "prompt_tokens",
          "total_tokens"
        ],
        "properties": {
          "prompt_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The number of tokens used by the prompt.",
            "minimum": 0
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "The total number of tokens used by the request.",
            "minimum": 0
          }
        }
      },
      "EmbeddingVector": {
        "oneOf": [
          {
            "type": "array",
            "items": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "type": "string"
          }
        ]
      },
      "EncodingFormat": {
        "type": "string",
        "enum": [
          "float",
          "base64"
        ]
      },
...
      "IcebergError": {
        "type": "object",
        "required": [
          "message",
          "type",
          "code"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/IcebergErrorType"
          }
        }
      },
      "IcebergErrorType": {
        "type": "string",
        "enum": [
          "NoSuchNamespaceException",
          "BadRequestException",
          "InternalServerError"
        ]
      },
      "IcebergResponseError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/IcebergError"
          }
        }
      },
...
      "Namespace": {
        "type": "object",
        "required": [
          "parts"
        ],
        "properties": {
          "parts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "NamespacesResponse": {
        "type": "object",
        "required": [
          "namespaces"
        ],
        "properties": {
          "namespaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Namespace"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}